UPD: move opt_beta method in conformal score class
